Fix tests in both the travis build and the jenkins build
