+++ docker/docker-compose.jenkins.yml
@@ -23,6 +23,6 @@ services:
   test:
     build:
       context: ..
-      dockerfile: Dockerfile
+      dockerfile: Dockerfile.spark
       target: metabrainz-spark-dev
-    command: dockerize -wait tcp://hadoop-master:9000 -timeout 60s bash -c "cp listenbrainz_spark/config.py.sample listenbrainz_spark/config.py; python -m pytest --junitxml=/data/test_report.xml --cov-report xml:/data/coverage.xml"
+    command: dockerize -wait tcp://hadoop-master:9000 -timeout 60s bash -c "cp listenbrainz_spark/config.py.sample listenbrainz_spark/config.py; python -m pytest -c pytest.spark.ini --junitxml=/data/test_report.xml --cov-report xml:/data/coverage.xml"
