+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-norecursedirs=listenbrainz/tests/integration
+norecursedirs=listenbrainz/tests/integration listenbrainz_spark
 addopts = --cov=listenbrainz
