+++ requirements_spark.txt
@@ -1,0 +1,12 @@
+click==7.0
+hetznercloud==1.1.1
+hdfs == 2.1.0
+Jinja2 == 2.10.1
+pika==0.13.0
+python-dateutil==2.8.0
+numpy==1.17.2
+git+https://github.com/metabrainz/brainzutils-python.git@v1.14.0
+pytest == 5.2.2
+pytest-cov == 2.8.1
+py4j == 0.10.8.1
+pyspark == 2.4.4
