+++ test.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-<<<<<<< HEAD
 # ./test.sh                bring up, make database, test, bring down
 # for development:
 # ./test.sh -u             bring up background and load database if needed
@@ -131,5 +130,3 @@ else
                    -p $COMPOSE_PROJECT_NAME \
                    run --rm listenbrainz py.test "$@"
 fi
-=======
->>>>>>> labs/master
